requirements.txt
django==6.0.3
django-allauth==65.15.0
django-crispy-forms==2.6
crispy-bootstrap5==2025.6
cryptography==46.0.5
oracledb==3.4.2
django-environ==0.13.0
pytest>=9.0.2
pytest-django>=4,<5
coverage[toml]>=7.10,<8
ruff>=0.13,<0.14
pre-commit>=4,<5
mypy>=1.18,<2
